+++ FluentNexus/Models/EndorsementRef.cs
@@ -1,11 +1,12 @@
 using System;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
 namespace Pathoschild.FluentNexus.Models
 {
     /// <summary>Simplified data about a mod endorsement.</summary>
-    public class EndorsementRef
+    public class EndorsementRef : IDataModel
     {
         /// <summary>The mod endorsement status.</summary>
         [JsonProperty("endorse_status")]
@@ -18,5 +19,9 @@ namespace Pathoschild.FluentNexus.Models
 
         /// <summary>The current mod version when the user endorsed the mod.</summary>
         public string Version { get; set; }
+
+        /// <inheritdoc />
+        [JsonExtensionData]
+        public IDictionary<string, object> UnmappedFields { get; set; }
     }
 }