+++ FluentNexus/Models/ModFileDownloadLink.cs
@@ -1,10 +1,11 @@
 using System;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace Pathoschild.FluentNexus.Models
 {
     /// <summary>A mod file download link.</summary>
-    public class ModFileDownloadLink
+    public class ModFileDownloadLink : IDataModel
     {
         /// <summary>The full name of the CDN serving the file.</summary>
         [JsonProperty("name")]
@@ -16,5 +17,9 @@ namespace Pathoschild.FluentNexus.Models
 
         /// <summary>The download URL.</summary>
         public Uri Uri { get; set; }
+
+        /// <inheritdoc />
+        [JsonExtensionData]
+        public IDictionary<string, object> UnmappedFields { get; set; }
     }
 }