+++ FluentNexus/Models/UserEndorsement.cs
@@ -1,11 +1,12 @@
 using System;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
 namespace Pathoschild.FluentNexus.Models
 {
     /// <summary>A mod endorsement by the user.</summary>
-    public class UserEndorsement
+    public class UserEndorsement : IDataModel
     {
         /// <summary>The unique mod ID.</summary>
         [JsonProperty("mod_id")]
@@ -24,5 +25,9 @@ namespace Pathoschild.FluentNexus.Models
 
         /// <summary>The mod endorsement status.</summary>
         public EndorsementStatus Status { get; set; }
+
+        /// <inheritdoc />
+        [JsonExtensionData]
+        public IDictionary<string, object> UnmappedFields { get; set; }
     }
 }