+++ FluentNexus/Models/UserRef.cs
@@ -1,9 +1,10 @@
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace Pathoschild.FluentNexus.Models
 {
     /// <summary>Simplified data about a user.</summary>
-    public class UserRef
+    public class UserRef : IDataModel
     {
         /// <summary>The unique user ID.</summary>
         [JsonProperty("member_id")]
@@ -15,5 +16,9 @@ namespace Pathoschild.FluentNexus.Models
 
         /// <summary>The username.</summary>
         public string Name { get; set; }
+
+        /// <inheritdoc />
+        [JsonExtensionData]
+        public IDictionary<string, object> UnmappedFields { get; set; }
     }
 }